Update CMakeLists and user guide
The CMakeLists now require that the path to the vhlle_params directory
is passed from the command line upon configuration. This way the user is
not required to copy the /eos directory by himself to the vhlle
directory. The README was updated accordingly.

Also the README now contains proper instructions regarding the
configuration of a soecific setup and the correct description of the
event-by-event runs and their averaging.