Update check_conservation script
The previous version of the script used to contain also the evaluation
of the evolution of the electric charge Q. In the latest version of
vHLLE however, the charge floating through the surface is however not
printed to the output. So, it cannot be properly analyzed. As a
consequence, everything related to extracting Q is removed from the
script to still have it functional.