Add RuRu and ZrZr hybrid setups
They are meant to rely on the specific initial nucleon distributions
provided by M. Alvioli. Be aware that these distribution files are not
shipped with the hybrid but need to be added manually to the configs
directory. Furthermore the corresponding paths in
configs/smash_initial_conditions_RuRu.yaml and
configs/smash_initial_conditions_ZrZr.yaml need to be updated
accordingly.

This branch is not intended to be merged in the current state