Remove all CMake related technology
This is the first commit moving towards the future scenario where a bash
handler will replace the CMake technology. As Git history retains older
code, there is no problem to radically remove whatever is going to be
replaced sooner or later. This should give a cleaner state from which to
start.