Implement missing test driver functions
The test driver is now ready for generalization. I implemented the
report and clean-up functionality and commented out not (yet) existing
functions, which will have to be provided on a per-case basis. As next
step, such an interface should be implemented and this should guarantee
that all functions invoked by the driver exist. It should be possible to
implement a function that takes the name of the function to be invoked
as first argument and possible further arguments afterwards. Using this
in the tests driver it would be clear which function has to be provided
by the interface.