Use associative arrays to globally declare valid keys
This allowed to stay completely general when parsing sections, although
some slightly involved indirection and some code duplication was needed.
However, the advantage of this approach makes it preferable, because any
further change to the code to add new configuration keys will be limited
to the place where global variables are declared.