Request more recent version of yq software
Since version 4.18.1 the eval action has been make the default one in
yq and with precedent versions such an action would have always been
needed. This would not be a big deal, but till version 4.11.0 feeding a
string into yq using process substitution <() was failing as yq was
attempting to modify that input (not permitted), resulting into a
  Error: seek /dev/fd/63: illegal seek
error. This has been fixed into version 4.11.1 and, given the extremely
simple way to download yq, we decided to bump the yq minimum
requirement to the more user friendly one.