Make logger use file descriptor 42
Using the file descriptor number 3 was of course fine, but usually fd 3
is the first free one and it is usually used for temporary redirection
of output. Hence, we use now 42 for the logger, so that collisions (and
possible confusion) are way less likely. Even if 42 is the default fd
one used by the logger when sourced without the --fd option, we still
use the option with 42 as value, so that we are unaffected by possible
future changes on the logger side.