Fix bug in system requirements when run in nounset mode
The auxiliary system information array uses strings with '|' as field
separator and word splitting on this character was used to assign an
array with the fields as entries. However, when fields are empty this
leads to access later not-existing array entries and this fails when
bash is in nounset mode. Now a '---' is used for all previously empty
fields.