Fix tput usage in system requirements
As tput relies on TERM environment variable which might be unset, we
decided to explicitly tell tput which terminal is used with the -T
option and passing ${TERM:-xterm} to it. However, it was not noticed
that the -T option makes tput not deduce the correct size of the
terminal and hence we do not have what we wish in the case when the
environment is set as it should. A new approach is used now.