Make code more robust for future new requirements
The function trying to find a program version needed to be changed for
each new possible requirement and, actually, was breaking in a very
confusing way when the needed change was forgotten (the global array was
not written to before returning). Now we made the default case to try
finding the version an internal error so that the developer understands
what is missing.