Avoid using system requirements when using them
The code became more complicated, but it makes sense to give a clean
report to the user and so we should not use system requirements that we
check for. Often a bash replacement is easy to implement. However, for
grep this might become too long and tricky and we decided for the moment
to work around this.