Implement mechanism to invoke codebase code in stricter mode
The hybrid handler bash code is run with some shell options enabled that
changes the shell behavior (nounset, errexit, etc.). It makes definitely
sense to run unit tests in the same mode. However, tests makes heavy use
of the $? variable to test for success/failure and this is the main
reason why in the beginning the shell standard behaviour had been
used. Now the same mode used by the handler is enabled in a function
that takes care to invoke a codebase function, catch and return its exit
code. Such as function will be next used in all unit test to run hybrid
handler snippets of code.