Refine mechanism to invoke codebase functions in unit tests
It can now be decided whether to invoke codebase functions in subshells
or not. Sometimes this is needed e.g. to avoid the test to be exited by
an exit command given in the invoked function, but this disable to
possibility to inspect changed values of global states (as the change
would occur in a subshell). Therefore, a mechanism to invoke codebase
functions without using a subshell has been added.