Treat auxiliary input file to Hydro as done in other blocks
In particular, we always want a symbolic link in the software output
folder to the auxiliary input file. This enables for easier
understanding of what was used for the run and hence makes future
reproducibility easier. If the user for some reason puts a file with the
right name in the output folder, this is used and no symlink is created.