Remove fancy printing from black-boxes and fix test
Python 3.12+ gives a warning when uncorrect escaped sequences are used
and this is annoying in our tests. I decided to remove the nice but
irrelevant fancy headers from the black-box mock script. I know it is
sad, but double escaping all the literal backslashes or using raw
strings and then fix endlines was not something I'd have preferred to
do...