Reformat full codebase with shfmt disabling -kp option
Actually the shfmt developers realised that the -kp has too many border
cases and, indeed, it does some weird formatting sometimes. They even
plan to remove this in the future v4 of their codebase. See issue
https://github.com/mvdan/sh/issues/658 for more information. Hence, we
decided to avoid using it here, too. After all, the result of the
formatting is not so bad looking and the main difference is that we now
have to only indent once more lines after the first of multi-line
commands instead of making them align after the command name of the
first line.