Install shfmt using go and picking latest version
I moved the installation of the custom bash-4.4 to the home as it was
done in the same folder of the codebase and this was making the
formatter also format bash files there and, hence, make the formatting
test fail. I actually think it is much better not to work in the
codebase folder to install anything.