Store scan objects while validating them
It makes sense to also store the scan object while validating them. The
validation requires quite some operations to go through the YAML
sections and most of this would be needed again to extract the scanning
information before producing all the possible combinations. It is
therefore handy to immediatelty store the validated information. This is
temporarily done into the associative array which will eventually
contain the list of values per parameter. This has the advantage that
that single array will be an in-out parameter for the functionality that
has to work out the values of each parameter.