Implement scan-prepare command line options parsing
The new execution mode has options in common with the already existing
do execution mode. The common options are parsed together in a common
parser function, but a validation of options per mode has been added and
performed before, such that it should not be possible to accept wrong
options. A mechanism to internally replace short options with long ones
and only deal with the latter has been introduced, too.