User older yq syntax not to have to bump version requirement
According to the yq release notes, the 'type' operator has been
introduced in version 4.25.1 as alias of the older 'tag' one. Since we
require version 4.18.1 using the new more descriptive alias would fail
on older accepted versions and it was now replaced by the older one.
Furthermore the yq short -r option was introduced in version 4.25.3 and
the longer --unwrapScalar should be preferred to stay compatible with
older versions. The releas notes can be found in the yq repository (at
https://github.com/mikefarah/yq/blob/master/release_notes.txt).