Fix function to ensure afterburner input consistency
Due to some missing $-signs a command expansion had become a simple
string and all checks to be performed were actually skipped. Once this
was fixed some unit tests failed and have been fixed, too (the
configuration file is now needed when testing the input preparation for
the afterburner). Furthermore some output message was split using double
quotes inside single quotes and this was not doing what is desired.