Make utilities interface more flexible and simplify code
The YAML utility functions consistently rely on the assumption that no
YAML key contains spaces or periods. Because of that, they were adjusted
to be callable either with single keys as arguments, or with all keys
concatenated with periods or a mix of both. Code that required a period
replacement and a split to stick to the old interface has been
simplified.