Bump system requirement for yq to version 4.24.2
This is the release of yq in which the comparison operators have been
introduced. Since yq is such a handy and powerful tool, in a project
already depending on it, I'd require a version that enables comparisons
as this make yq handy in the shell to be used for floating point
comparisons. Its installation is so trivial that bumping this
requirement is not a problem at all.