Give preference to run ID specified on command line option
If the run ID is specified both on the command line option and in the
configuration file, the value provided on the command line is used. To
avoid confusion, the value of the run ID in the configuration is changed
when extracting the sections to store in the output folder.