Fix unit test framework about calling codebase functions
This was not done as it should have been. The functions of the codebase
must be called after having enabled the bash behaviour used by the
handler. This was done, but then the function was called with a trailing
"|| return_code=$?" and this made the l.h.s. of the logical or ignore
the errexit mode. This is fixed now.