Split check about needed Python requirements
When system requirements are checked it is only possible to test for
always needed Python requirements, while checks that depend on user
input (like if LHS is used in prepare-scan mode) can only be done after
the handler configuration has been parsed. The latter have been moved to
the sanity checks now.