Refine exit code of utility function to check existence
If files are tested to ensure existence, then the exit code should
reflect that they have not been found in case of failure. In the
opposite case, when files are expected to not exist but they do, then a
logic exit code should be used. This enables better testing in other
contexts.