Generalise SMASH-hadron-sampler black-box script
Now the mock script can mimic both older and newer SMASH-sampler
behaviour and, in particular, has both interfaces and is able to deal
with both old and new configuration files. The version and hence the
behaviour of the script can be set using an environment variable, which
is defaulted to the more recent version.