Extract sampler changing input keys into global array
To avoid spreading around the test on the sampler version, I decided to
introduce a new constant global associative array to store the names of
the configuration keys that changed into version 3.2 into variables. In
tests it is still needed to do the logic by hand because we want to
indeed test the validation mechanism without using the codebase logic.