Implement procedure to print faulty keys in YAML replacement
While adding this feature, I realised that the previous approach (to
understand wheather the YAML keys replacement was possible or not) was
not always working. In particular, the basic idea was to count how many
lines the initial YAML file had, merge it with the new YAML keys and
fail if the resulting number of lines was different. However, changing
the length of an array was not correctly taken into account (this in
general changes the number of lines). Note that wrapping arrays using
the [...] notation is not in general a good solution, as arrays might
contain maps and be arbitrarily complex.

Now the keys to be replaced are parsed into a much more structured way
and the same parsing is done on the file. In this way two lists are
created that are compared. If any of the keys to be substituted is not
present in the YAML file, an error is given and a report can also be
provided to the user.