Restrict `Add_spectators_from_IC: true` key to one IC event
Currently, we only support the `Add_spectators_from_IC` key set to true
when only one IC event is run. This commit implements a restriction
regarding this combination in the afterburner stage and adds a test to
check the afterburner functionality concerning this behavior.
IMPORTANT: This makes the code fail only in the afterburner stage when
the python script add_spectators.py tries to add the spectators from the
IC particle list to the Sampler particle list.