Rename sampler base configuration files
The new names are consistent with those of the IC ones. Having the same
notation for all makes sense and a more compact notation is to be
preferred. The name is in any case internal and in real runs the file is
copied to the run folder using at destination a different more
user-friendly name.